Update custom parameter

+++ include/phase_metrics.h
@@ -77,7 +77,7 @@ public:
     int64_t bytes_per_second() const noexcept;
 
     //! Get custom integers map
-    const std::map<std::string, int>& custom_int() const noexcept { return _custom_int; }
+    const std::map<std::string, int64_t>& custom_int() const noexcept { return _custom_int; }
     //! Get custom strings map
     const std::map<std::string, std::string>& custom_str() const noexcept { return _custom_str; }
 
@@ -105,7 +105,7 @@ public:
         \param name - Name
         \param value - Value
     */
-    void SetCustom(const std::string& name, int value)
+    void SetCustom(const std::string& name, int64_t value)
     { _custom_int[name] = value; }
     //! Set custom string value
     /*!
@@ -122,7 +122,7 @@ private:
     int64_t _total_iterations;
     int64_t _total_items;
     int64_t _total_bytes;
-    std::map<std::string, int> _custom_int;
+    std::map<std::string, int64_t> _custom_int;
     std::map<std::string, std::string> _custom_str;
 
     int64_t _iterstamp;
